Prototype of 0.13upgrade changes for redirect
If a provider changes namespace in the registry, we can detect this when
running the 0.13upgrade command. As long as there is a version matching
the user's constraints, we should use the provider's new source address.
Otherwise, notify the user that the provider has moved and a version
upgrade is necessary to move to it.

Untested, no test coverage, and the diags could probably use some work.